requirements.txt
PySide6>=6.9.0
SQLAlchemy>=2.0.23
PyMySQL>=1.1.0
bcrypt>=4.1.1
cryptography>=41.0.7
alembic>=1.13.1
python-dotenv>=1.0.0
